Factor temporary logic out into a separate module
