+++ server/src/lib.rs
@@ -13,7 +13,7 @@ use {
         primitives::U256,
         state_actor::StatePayloadId,
         storage::InMemoryState,
-        types::state::{Command, RpcBlock, StateMessage},
+        types::state::{Command, StateMessage},
     },
     once_cell::sync::Lazy,
     std::{
@@ -35,6 +35,7 @@ use {
     },
 };
 
+mod geth_genesis;
 mod mirror;
 
 #[cfg(test)]
@@ -239,9 +240,9 @@ async fn mirror(
 
     // TODO: this is a hack because we currently can't compute the genesis
     // hash expected by op-node.
-    if is_genesis_block_request(&request).unwrap_or(false) {
-        let block =
-            extract_genesis_block(&parsed_geth_response).expect("Must get genesis from geth");
+    if geth_genesis::is_genesis_block_request(&request).unwrap_or(false) {
+        let block = geth_genesis::extract_genesis_block(&parsed_geth_response)
+            .expect("Must get genesis from geth");
         state_channel
             .send(Command::GenesisUpdate { block }.into())
             .await
@@ -252,47 +253,6 @@ async fn mirror(
 
     let body = hyper::Body::from(serde_json::to_vec(&op_move_response).unwrap());
     Ok(warp::reply::Response::new(body))
-}
-
-fn is_genesis_block_request(request: &serde_json::Value) -> Option<bool> {
-    let obj = request.as_object()?;
-    let method = obj.get("method")?.as_str()?;
-    if method != "eth_getBlockByNumber" {
-        return Some(false);
-    }
-    let first_param = obj.get("params")?.as_array()?.first()?.as_str()?;
-    Some(first_param == "0x0")
-}
-
-fn extract_genesis_block(geth_response: &serde_json::Value) -> Option<ExtendedBlock> {
-    let geth_block: RpcBlock =
-        serde_json::from_value(geth_response.as_object()?.get("result")?.clone()).ok()?;
-    let header = Header {
-        parent_hash: geth_block.header.parent_hash,
-        ommers_hash: geth_block.header.ommers_hash,
-        beneficiary: geth_block.header.beneficiary,
-        state_root: geth_block.header.state_root,
-        transactions_root: geth_block.header.transactions_root,
-        receipts_root: geth_block.header.receipts_root,
-        logs_bloom: geth_block.header.logs_bloom,
-        difficulty: geth_block.header.difficulty,
-        number: geth_block.header.number,
-        gas_limit: geth_block.header.gas_limit,
-        gas_used: geth_block.header.gas_used,
-        timestamp: geth_block.header.timestamp,
-        extra_data: geth_block.header.extra_data.clone(),
-        nonce: geth_block.header.nonce,
-        base_fee_per_gas: geth_block.header.base_fee_per_gas,
-        withdrawals_root: geth_block.header.withdrawals_root,
-        blob_gas_used: geth_block.header.blob_gas_used,
-        excess_blob_gas: geth_block.header.excess_blob_gas,
-        parent_beacon_block_root: geth_block.header.parent_beacon_block_root,
-        mix_hash: geth_block.header.mix_hash,
-        requests_hash: geth_block.header.requests_hash,
-    };
-    let block = Block::new(header, Vec::new());
-    let ext_block = ExtendedBlock::new(geth_block.header.hash, Default::default(), block);
-    Some(ext_block)
 }
 
 async fn proxy(
